Script to check ratio text/html
It's thought to be run locally. If you're interested in running it
take a look at my crontab file as an example:

0 * * * * /home/pepellou/Workspace/crisis-what-crisis/script/check_text_html.bash

In the script you have a log_file var with the path to your desired
log file. Remember to set permissions to your user to write that file.